frontend/refactor: ChatBubble / SystemMessage
