+++ wetty-chat-mobile/src/components/chat/MessageOverlay.module.scss
@@ -50,143 +50,8 @@
   align-items: flex-end;
 }
 
-/* Bubble clone mirrors ChatBubble and only enables selection after mouse detection */
 .bubbleClone {
-  padding: 8px 12px;
-  line-height: 1.4;
-  font-size: 14px;
-  user-select: none;
-  -webkit-user-select: none;
-  -webkit-touch-callout: none;
   box-shadow: 0 8px 32px rgba(0, 0, 0, 0.28);
-}
-
-.mouseSelectable {
-  user-select: text;
-  -webkit-user-select: text;
-  -webkit-touch-callout: default;
-}
-
-.bubbleReceived {
-  background: #f0f0f0;
-  color: #1a1a1a;
-  border-radius: 18px 18px 18px 4px;
-}
-
-.bubbleSent {
-  background: #3b82f6;
-  color: #fff;
-  border-radius: 18px 18px 4px 18px;
-}
-
-.senderName {
-  font-size: 12px;
-  font-weight: 600;
-  margin-bottom: 2px;
-  opacity: 0.85;
-}
-
-.replyPreview {
-  border-left: 3px solid rgba(255, 255, 255, 0.5);
-  background: rgba(0, 0, 0, 0.1);
-  padding: 4px 8px;
-  border-radius: 6px;
-  margin-bottom: 6px;
-}
-
-.bubbleReceived .replyPreview {
-  border-left-color: #3b82f6;
-  background: rgba(0, 0, 0, 0.06);
-}
-
-.replyPreviewName {
-  font-size: 11px;
-  font-weight: 600;
-  opacity: 0.85;
-}
-
-.replyPreviewText {
-  font-size: 12px;
-  opacity: 0.7;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.messageWrapper {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: flex-end;
-}
-
-.messageText {
-  word-break: break-word;
-  white-space: pre-wrap;
-}
-
-.messageLink {
-  color: inherit;
-  text-decoration: underline;
-}
-
-.timestampSpacer {
-  flex-grow: 1;
-}
-
-.timestamp {
-  font-size: 12px;
-  opacity: 0.7;
-  margin-left: 1.5ch;
-  display: inline-flex;
-  align-items: center;
-  gap: 4px;
-}
-
-.statusIcon {
-  font-size: 14px;
-}
-
-.attachmentsContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: end;
-  gap: 8px;
-  margin: 4px 0;
-}
-
-.attachmentImage {
-  display: block;
-  max-width: 100%;
-  max-height: 200px;
-  border-radius: 8px;
-  object-fit: contain;
-}
-
-.filePlaceholder {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 10px 12px;
-  border-radius: 8px;
-  background: rgba(0, 0, 0, 0.06);
-  text-decoration: none;
-  color: inherit;
-}
-
-.bubbleSent .filePlaceholder {
-  background: rgba(255, 255, 255, 0.15);
-}
-
-.fileIcon {
-  font-size: 22px;
-  flex-shrink: 0;
-  opacity: 0.7;
-}
-
-.fileName {
-  font-size: 13px;
-  word-break: break-all;
-  opacity: 0.9;
 }
 
 /* Reaction bar */
