+++ wetty-chat-mobile/src/components/chat/messages/MessageDateSeparator.module.scss
@@ -1,0 +1,15 @@
+.separator {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 12px 0 8px;
+}
+
+.label {
+  background-color: rgba(var(--ion-text-color-rgb, 0, 0, 0), 0.08);
+  color: var(--ion-color-medium, #555);
+  font-size: 13px;
+  font-weight: 500;
+  padding: 4px 12px;
+  border-radius: 12px;
+}
