+++ wetty-chat-mobile/src/components/chat/messages/MessageDateSeparator.tsx
@@ -1,0 +1,13 @@
+import styles from './MessageDateSeparator.module.scss';
+
+interface MessageDateSeparatorProps {
+  label: string;
+}
+
+export function MessageDateSeparator({ label }: MessageDateSeparatorProps) {
+  return (
+    <div className={styles.separator}>
+      <span className={styles.label}>{label}</span>
+    </div>
+  );
+}
