+++ wetty-chat-mobile/src/components/chat/messages/SystemMessage.module.scss
@@ -1,0 +1,15 @@
+.container {
+  display: flex;
+  justify-content: center;
+  padding: 8px 16px;
+}
+
+.content {
+  max-width: min(100%, 520px);
+  color: var(--ion-color-medium, #555);
+  font-size: 13px;
+  line-height: 1.45;
+  text-align: center;
+  white-space: pre-wrap;
+  word-break: break-word;
+}
