+++ wetty-chat-mobile/src/components/chat/messages/SystemMessage.tsx
@@ -1,0 +1,13 @@
+import styles from './SystemMessage.module.scss';
+
+interface SystemMessageProps {
+  message: string;
+}
+
+export function SystemMessage({ message }: SystemMessageProps) {
+  return (
+    <div className={styles.container}>
+      <div className={styles.content}>{message}</div>
+    </div>
+  );
+}
