+++ wetty-chat-mobile/src/pages/chat-thread/chat-thread.scss
@@ -75,23 +75,6 @@
   cursor: pointer;
 }
 
-/* Message rows */
-.chat-date-separator {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 12px 0 8px;
-
-  span {
-    background-color: rgba(var(--ion-text-color-rgb, 0, 0, 0), 0.08);
-    color: var(--ion-color-medium, #555);
-    font-size: 13px;
-    font-weight: 500;
-    padding: 4px 12px;
-    border-radius: 12px;
-  }
-}
-
 .message-row {
   display: flex;
   align-items: flex-end;
